+++ qmt/functions/oriest.py
@@ -25,7 +25,8 @@ def oriEstVQF(gyr, acc, mag=None, params=None, debug=False, plot=False):
     """
     VQF orientation estimation algorithm.
 
-    See https://arxiv.org/abs/2203.17024 and https://github.com/dlaidig/vqf for more information about this algorithm.
+    See https://github.com/dlaidig/vqf, https://doi.org/10.1016/j.inffus.2022.10.014, and
+    https://arxiv.org/abs/2203.17024 for more information about this algorithm.
 
     If potential real-time capability is not needed, use the offline version :meth:`qmt.oriEstOfflineVQF` for improved
     accuray. This algorithm is also available as an online data processing block: :class:`qmt.OriEstVQFBlock`.
@@ -113,7 +114,8 @@ def oriEstBasicVQF(gyr, acc, mag=None, params=None, debug=False, plot=False):
     Basic version of the VQF orientation estimation algorithm (no rest detection, no gyroscope bias estimation, no
     magnetic disturbance rejection).
 
-    See https://arxiv.org/abs/2203.17024 and https://github.com/dlaidig/vqf for more information about this algorithm.
+    See https://github.com/dlaidig/vqf, https://doi.org/10.1016/j.inffus.2022.10.014, and
+    https://arxiv.org/abs/2203.17024 for more information about this algorithm.
 
     :param gyr: Nx3 array with gyroscope measurements [rad/s]
     :param acc: Nx3 array with accelerometer measurements [m/s^2]
@@ -186,7 +188,8 @@ def oriEstOfflineVQF(gyr, acc, mag=None, params=None, debug=False, plot=False):
     """
     Offline version of the VQF orientation estimation algorithm.
 
-    See https://arxiv.org/abs/2203.17024 and https://github.com/dlaidig/vqf for more information about this algorithm.
+    See https://github.com/dlaidig/vqf, https://doi.org/10.1016/j.inffus.2022.10.014, and
+    https://arxiv.org/abs/2203.17024 for more information about this algorithm.
 
     :param gyr: Nx3 array with gyroscope measurements [rad/s]
     :param acc: Nx3 array with accelerometer measurements [m/s^2]