+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.3] - 2022-11-01
 ### Added
 - Ability to use custom IMU boxes in BoxModel and UiIMUScene.
 - Option to rotate camera in UiIMUScene on click.
@@ -69,7 +71,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/dlaidig/qmt/compare/v0.2.2...HEAD
+[Unreleased]: https://github.com/dlaidig/qmt/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/dlaidig/qmt/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/dlaidig/qmt/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/dlaidig/qmt/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/dlaidig/qmt/compare/v0.1.0...v0.2.0